reorganize vault related files and define messages
