+++ src/lib.rs
@@ -119,6 +119,7 @@ pub mod cli;
 pub mod command;
 pub mod errors;
 pub mod file;
+pub mod message;
 pub mod operation;
 pub mod output;
 pub mod reads;
@@ -127,4 +128,5 @@ pub mod secure;
 pub mod store;
 pub mod utils;
 pub mod vault;
-pub mod vault_encrypted;
+
+pub type Password = String;
