+++ src/vault/mod.rs
@@ -1,3 +1,6 @@
+pub mod encrypted;
+pub mod interface;
+
 use core::str;
 use std::collections::HashMap;
